Fix `version` for AR 4.0.x
